Use `getunsolvedblock` RPC for mining
This RPC serializes the whole block as a hex string into JSON. It also
constructs the coinbase transaction. Because of the larger nonces in
Lotus, the mining software does not need to touch the merkle tree
whatsoever. This is nice because the merkle tree construction has
changed, and the current code does not work against the latest Lotus
blocks.